Expect.(notE|e)qual fails if given floats
This is code for 230 (hopefully GitHub doesn't pick this up because stealth).

I tried Richard's proposed implementation and it had an issue with a
large int value in the seed tests. I think this way is more robust: a
value is a float if its string version parses as a float but not as an
int.