Refactor the running of fuzz tests
This separates concerns into different functions, provides some better
names, and deletes out-of-date comments, without changing the signature
of the exposed function.

Hopefully this makes the algorithm much less "magic" and "Max's domain
knowledge". This PR is a good opportunity for some other folks to become
familiar with this code.