Remove Simplify.map and Simplify.andMap
These functions operate on LazyList, which we are hiding (so this
finishes the work mostly done in #79). Grepping for "lazy" in the
generated docs JSON finds no matches now, so I think there are not any
more that I've missed.

It's impossible to implement these functions (sensibly) if they were to
accept simplifiers instead of lazy lists. A simplifier is a function,
and a value of the type of the argument isn't available.

By deleting the relevant documentation, this closes #65.

Tagging major release blocker for obvious reasons.