Swap the current rose-tree approach for "integrated shrinking" one
Based on Hypothesis [1] / Minithesis [2] / elm-microthesis [3].

Instead of simplifying generated values, we add a layer of indirection: remember
the PRNG drawn numbers (called `RandomRun`, can be thought of as a `List Int`)
from which we generated the values, simplify the RandomRun irrespective of the
values or fuzzers used, and then try generating values from the shrunk RandomRun
again. Fuzzers are constructed in such a way to nicely synergize with the
simplifying strategies (called `ShrinkCmd`).

Explained in more detail in:

- Hypothesis blogpost "Integrated vs type based shrinking" [4]
- Minithesis repo [2]
- Introductory video to the "Writing a property-based testing library in Elm" YouTube series [5]

[1] https://github.com/HypothesisWorks/hypothesis
[2] https://github.com/DRMacIver/minithesis
[3] https://github.com/Janiczek/elm-microthesis
[4] https://hypothesis.works/articles/integrated-shrinking/
[5] https://www.youtube.com/watch?v=Pym32n6AfSs&list=PLymbvEbZ-wpXqyoqFpmUYiNcy2E7r3vzH&index=1