Increase RandomRun max length
In elm-in-elm/compiler, in an Expr fuzzer of max depth 3 with lists
and strings hardcoded to length 5 this was maxing out.

Now of course the sensible solution was to make the length variable
(1-5 etc). but this anecdote means the limit can be reached by more
advanced fuzzers.

So let's raise the limit a bit? List of 64k integers still seems OK
memory-hungriness-wise.